Implement the backup.
The code creates the dump in the directory named for the finest-grained
interval specified. The dump file uses an ISO-8601 UTC timestamp for the file
name. The tests make sure this all works, and include a mock `pg_dump` program
that echos its arguments to the dump file, both so the dump file is created,
and so that we can verify that the proper options have been passed to
`pg_dump`.

Next up:

* Iterating over intervals to keep and linking the dump file as appropriate.
* Cleaning up older dump files.