Reorganize things a bit.
Don't have `_link_for()` call `_needs_link()`; have `_rolldump()` call them
both. `_link_for()` now returns the destination file name so that it can be
added to the list of files for the given interval.

While at it, Change the `_files_for()` test to use "days" instead of "hours"
so that later tests that deal with hours won't notice them.